requirements.txt
asgiref==3.3.1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.0.3
Django==3.1.7
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-split-settings==1.0.1
django-storages==1.11.1
djangorestframework==3.12.2
drf-yasg==1.20.0
ecdsa==0.14.1
firebase-admin==4.5.3
git+https://github.com/dimsamaras/soft-delete.git#egg=soft-delete
git+https://github.com/dimsamaras/bubbling_firebase_authentication.git#egg=bubbling-firebase-authentication
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
httplib2==0.19.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msgpack==1.0.2
packaging==20.9
proto-plus==1.18.0
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.15.0
python-jose==3.2.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
